rename MethodNameReplacerRector to RenameMethodRector
