--- tests/Rector/MethodCall/MethodNameReplacerRector/Source/Something.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\MethodCall\MethodNameReplacerRector\Source;
-
-final class Something
-{
-
-}
