+++ tests/Rector/MethodCall/RenameMethodRector/Fixture/fixture2.php.inc
